pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "setlistfm-playlist-sync"
version = "0.1.0"
description = "Create YouTube Music playlists from Setlist.fm data"
readme = "README.md"
authors = [{name = "anatoli12", email = ""}]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio",
]
keywords = ["music", "youtube", "setlist.fm", "playlist"]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.25.0",
    "ytmusicapi>=1.0.0",
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]

[project.urls]
Homepage = "https://github.com/anatoli12/setlistfm-playlist-sync"
Repository = "https://github.com/anatoli12/setlistfm-playlist-sync.git"
Issues = "https://github.com/anatoli12/setlistfm-playlist-sync/issues"

[project.scripts]
setlistfm-playlist-sync = "setlistfm_playlist_sync.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=setlistfm_playlist_sync --cov-report=term-missing"
